# support/openapi/api-docs.yaml
openapi: 3.0.1
info:
  title: Email Simualtor API
  description: Spring Boot REST API for email simualtor
  termsOfService: https://github.com/lzugaj/email-task-server
  contact:
    name: Luka Zugaj
    url: https://github.com/lzugaj
    email: luka.zugaj7@gmail.com
  license:
    name: Apache License Version 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Email Message
  description: Email message API
paths:
  /api/v1/email-message:
    get:
      tags:
      - Email Message
      summary: Find all email messages
      operationId: findAll
      responses:
        "200":
          description: Found all email messages
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmailMessageResponse'
    post:
      tags:
      - Email Message
      summary: Send a new email message
      operationId: send
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailMessageRequest'
        required: true
      responses:
        "400":
          description: Invalid input
          content:
            '*/*':
              schema:
                type: string
                enum:
                - 100 CONTINUE
                - 101 SWITCHING_PROTOCOLS
                - 102 PROCESSING
                - 103 CHECKPOINT
                - 200 OK
                - 201 CREATED
                - 202 ACCEPTED
                - 203 NON_AUTHORITATIVE_INFORMATION
                - 204 NO_CONTENT
                - 205 RESET_CONTENT
                - 206 PARTIAL_CONTENT
                - 207 MULTI_STATUS
                - 208 ALREADY_REPORTED
                - 226 IM_USED
                - 300 MULTIPLE_CHOICES
                - 301 MOVED_PERMANENTLY
                - 302 FOUND
                - 302 MOVED_TEMPORARILY
                - 303 SEE_OTHER
                - 304 NOT_MODIFIED
                - 305 USE_PROXY
                - 307 TEMPORARY_REDIRECT
                - 308 PERMANENT_REDIRECT
                - 400 BAD_REQUEST
                - 401 UNAUTHORIZED
                - 402 PAYMENT_REQUIRED
                - 403 FORBIDDEN
                - 404 NOT_FOUND
                - 405 METHOD_NOT_ALLOWED
                - 406 NOT_ACCEPTABLE
                - 407 PROXY_AUTHENTICATION_REQUIRED
                - 408 REQUEST_TIMEOUT
                - 409 CONFLICT
                - 410 GONE
                - 411 LENGTH_REQUIRED
                - 412 PRECONDITION_FAILED
                - 413 PAYLOAD_TOO_LARGE
                - 413 REQUEST_ENTITY_TOO_LARGE
                - 414 URI_TOO_LONG
                - 414 REQUEST_URI_TOO_LONG
                - 415 UNSUPPORTED_MEDIA_TYPE
                - 416 REQUESTED_RANGE_NOT_SATISFIABLE
                - 417 EXPECTATION_FAILED
                - 418 I_AM_A_TEAPOT
                - 419 INSUFFICIENT_SPACE_ON_RESOURCE
                - 420 METHOD_FAILURE
                - 421 DESTINATION_LOCKED
                - 422 UNPROCESSABLE_ENTITY
                - 423 LOCKED
                - 424 FAILED_DEPENDENCY
                - 425 TOO_EARLY
                - 426 UPGRADE_REQUIRED
                - 428 PRECONDITION_REQUIRED
                - 429 TOO_MANY_REQUESTS
                - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
                - 451 UNAVAILABLE_FOR_LEGAL_REASONS
                - 500 INTERNAL_SERVER_ERROR
                - 501 NOT_IMPLEMENTED
                - 502 BAD_GATEWAY
                - 503 SERVICE_UNAVAILABLE
                - 504 GATEWAY_TIMEOUT
                - 505 HTTP_VERSION_NOT_SUPPORTED
                - 506 VARIANT_ALSO_NEGOTIATES
                - 507 INSUFFICIENT_STORAGE
                - 508 LOOP_DETECTED
                - 509 BANDWIDTH_LIMIT_EXCEEDED
                - 510 NOT_EXTENDED
                - 511 NETWORK_AUTHENTICATION_REQUIRED
        "201":
          description: Email message sent
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EmailMessageRequest'
  /api/v1/email-message/{id}:
    get:
      tags:
      - Email Message
      summary: Find a email message by id
      operationId: findById
      parameters:
      - name: id
        in: path
        description: Id of the email message to be founded. Cannot be empty.
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found the email message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailMessageResponse'
        "404":
          description: Book not found
        "400":
          description: Invalid id supplied
components:
  schemas:
    EmailMessageRequest:
      required:
      - from
      - importance
      - subject
      - to
      type: object
      properties:
        from:
          type: string
          description: Who is sending the email
          example: john.doe@gmail.com
        to:
          type: string
          description: To whom is mail sending
          example: john.doe@gmail.com
        cc:
          type: string
          description: Who also needs a copy of this email
          example: "john.doe@gmail.com,spring.boot@gmail.com"
        subject:
          type: string
          description: Subject of the email
          example: This is the subject of the email
        importance:
          type: string
          description: Choose an importance of the email
          example: HIGH
        content:
          type: string
          description: Content of the email
          example: This is the content of the email
    EmailMessageResponse:
      type: object
      properties:
        id:
          type: integer
          description: Id of the email
          format: int64
          example: 234
        from:
          type: string
          description: Who is sending the email
          example: john.doe@gmail.com
        to:
          type: string
          description: To whom is mail sending
          example: john.doe@gmail.com
        cc:
          type: string
          description: Who also needs a copy of this email
          example: "john.doe@gmail.com,spring.boot@gmail.com"
        subject:
          type: string
          description: Subject of the email
          example: This is the subject of the email
        importance:
          type: string
          description: Importance of the email
          example: HIGH
        content:
          type: string
          description: Content of the email
          example: This is the content of the email